requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain_text_splitters
langchain_openai
chromadb
langchain-community
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
beautifulsoup4
streamlit